requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autopep8==2.0.4
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.10
django-silk==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
gprof2dot==2022.7.29
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
